requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=2.0.0,<2.1.0  # Constraint for numba compatibility (numba requires numpy<2.1)
scikit-learn>=1.3.0
scipy>=1.10.0

# Machine Learning Models
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Explainability
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Framework
Flask>=2.3.0
flask-cors>=4.0.0

# PDF Generation
reportlab>=4.0.0

# Utilities
joblib>=1.3.0
openpyxl>=3.1.0

# Imbalanced Learning
imbalanced-learn>=0.11.0

# Optional: For PDF text extraction (if needed)
# PyPDF2>=3.0.0

# Chatbot
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
groq>=0.4.0

# Note: Dependency Warnings
# If you have gensim or tensorflow installed system-wide (not in this project's requirements),
# you may see dependency warnings. These packages are not required for this project and the 
# warnings can be safely ignored. The project uses:
# - numpy>=2.0.0,<2.1.0 (for SHAP and numba compatibility)
# - protobuf>=6.0.0 (latest version)
# 
# System-wide packages like gensim (requires numpy<2.0) and tensorflow (requires protobuf<6.0)
# will show warnings but won't affect this project's functionality.

